Implement transposition table update operations and utility functions
- Added a new module for transposition table update operations, including a generic helper function for updating entries with depth filtering using compare-and-swap (CAS).
- Introduced utility functions for the transposition table, including depth extraction and depth threshold calculation.
- Updated the transposition table trait implementation to use the new entry structure.
- Modified benchmark tests to utilize the new methods for probing and updating transposition table entries.
- Refactored garbage collection and prefetching benchmarks to align with the new API changes.
- Enhanced metrics collection during CAS operations for better performance tracking.
