+++ packages/rust-core/crates/engine-core/src/search/tt/prefetch.rs
@@ -1,0 +1,6 @@
+//! Prefetch functionality for transposition table
+//!
+//! This module provides adaptive prefetching with integration to the
+//! existing AdaptivePrefetcher implementation.
+
+pub use crate::search::adaptive_prefetcher::{AdaptivePrefetcher, PrefetchStats};
