+++ packages/rust-core/crates/tools/src/bin/tt_cas_update_bench.rs
@@ -121,7 +121,7 @@ fn run_benchmark(config: &BenchConfig) -> (Duration, u64, Option<DetailedMetrics
     let total_stores = stats.total_stores.load(Ordering::Relaxed);
 
     // Get metrics
-    let metrics = tt.metrics.as_ref().map(|m| DetailedMetrics {
+    let metrics = tt.metrics().as_ref().map(|m| DetailedMetrics {
         cas_attempts: m.cas_attempts.load(Ordering::Relaxed),
         cas_successes: m.cas_successes.load(Ordering::Relaxed),
         cas_failures: m.cas_failures.load(Ordering::Relaxed),
